🚂Move components to specified folders
