+++ src/app/app-routing.module.ts
@@ -1,15 +1,17 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { LandingPageComponent } from './components/landing-page/landing-page.component';
-import  { SigninComponent } from './components/signin/signin.component';
+import { LandingPageComponent } from './pages/landing-page/landing-page.component';
+import { SigninComponent } from './components/signin/signin.component';
 import { ErrorPageComponent } from './pages/error-page/error-page.component';
+import { SignUpComponent } from './components/sign-up/sign-up.component';
 
 //Main routes module
 const routes: Routes = [
   { path: 'home', component: LandingPageComponent },
   { path: 'sign-in', component: SigninComponent },
-  { path: '',   redirectTo: '/home', pathMatch: 'full' },
-  { path : '**', component: ErrorPageComponent }
+  { path: 'register-user', component: SignUpComponent },
+  { path: '', redirectTo: '/home', pathMatch: 'full' },
+  { path: '**', component: ErrorPageComponent }
 ];
 
 @NgModule({
