--- src/app/components/landing-page/landing-page.component.html
@@ -1,1 +1,0 @@
-<p>landing-page works!</p>
