+++ src/app/components/sign-up/sign-up.component.scss
@@ -1,0 +1,88 @@
+
+mat-card{
+    padding-top: 30px;
+    height: 600px;
+    width: 600px;
+    padding-bottom: 50px;
+}
+
+.loginPic {
+    height: 500px;
+    width: 500px;
+}
+
+.submit {
+    width: 300px;
+    border-radius: 12px;
+    font-size: 16px;
+}
+
+.form-label {
+    font-size: 16px;
+}
+
+.auth-card {
+	border-radius: 8px;
+	background: white;
+	box-shadow: 0px 3px 6px #70739D0B;
+	border: 1px solid #fdfdfd;
+	border-radius: 8px;
+}
+ 
+mat-card {
+    margin-left: 100px;
+}
+
+.salutation {
+    width: 400px;
+    font-size: 30px;
+}
+
+.p1 {
+    color: #54626F
+}
+
+.login-btns {
+    padding-top: 15px;
+}
+
+.btn-primary {
+    color:#7532FA;
+}
+
+.google {
+    padding-top: 20px;
+}
+
+.btn-secondary {
+    font-weight: bold;
+    font-family: 'Roboto';
+    font-size: 14px;
+    border: none;
+    background-color: white;
+    color: #7532FA ;
+    font-weight: bold;
+    width: 100%;
+
+  }
+
+.btn-primary{
+    font-weight: bold;
+    font-family: 'Roboto';
+    font-size: 14px;
+    border: none;
+    background-color: white;
+    color: #7532FA ;
+    font-weight: bold;
+    width: 100%;
+
+}
+
+.page-gif {
+    width: 60px;
+    padding-bottom: 20px;
+}
+
+.google-btn{
+    width: 20px;
+}
