+++ src/app/pages/landing-page/landing-page.component.html
@@ -1,0 +1,18 @@
+<ng-container>
+  <div class="page-content" fxLayout="column" fxlayoutAlign="space-between center" fxLayoutGap="15px">
+    <span class="title">Wizardly Persona</span>
+    <section class="card-section" fxLayout="row" fxLayoutAlign="space-between center" fxLayoutGap="15px">
+      <mat-card>
+        Conent 1
+      </mat-card>
+
+      <mat-card>
+        Conent 1
+      </mat-card>
+
+      <mat-card>
+        Conent 1
+      </mat-card>
+    </section>
+  </div>
+</ng-container>
