+++ src/app/pages/landing-page/landing-page.component.scss
@@ -1,0 +1,3 @@
+.page-content{
+  padding: 10%;
+}
